Some cleanup around ConfigureSettingsBuildOperationType
- extract ConfigureSettingsBuildOperationType.ProjectDescription into interface
- add proper javadoc for ConfigureSettingsBuildOperationType.ProjectDescription
- guarantee order of entries in ConfigureSettingsBuildOperationType.ProjectDescription#Children
